Makes default delete action check for post method.
Updates admin theme to use buttons instead of links for deletion.

+++ application/libraries/Omeka/Controller/Action.php
@@ -501,7 +501,12 @@ abstract class Omeka_Controller_Action extends Zend_Controller_Action
      * @return void
      */
     public function deleteAction()
-    {        
+    {
+        if (!$this->getRequest()->isPost()) {
+            $this->_forward('error');
+            return;
+        }
+
         $record = $this->findById();         
         $record->delete();
